implement workorder BATCH-UPDATE

+++ schema/jcmi_hms_procs.sql
@@ -1032,24 +1032,26 @@ ELSEIF (_action = 'PARTIALTIAL_UPDATE') THEN
 		EXECUTE stmt;
 		DEALLOCATE PREPARE stmt;
 
-ELSEIF(_action ='UPDATE') THEN
+ELSEIF(_action ='UPDATE' OR _action = 'BATCH-UPDATE') THEN
 		IF (_workorderUUID IS NULL) THEN
 			SIGNAL SQLSTATE '45001' SET MESSAGE_TEXT = 'call WORKORDER_workOrder: _workorderUUID is null for UPDATE action';
 			LEAVE WORKORDER_workOrder;
 		END IF;
 
-        select workorder_tag INTO _workorder_tag where workorderUUID = _workorderUUID;
+        IF (_action = 'UPDATE') THEN
+            select workorder_tag INTO _workorder_tag where workorderUUID = _workorderUUID;
 
-        IF (_workorder_tag IS NOT NULL) THEN
-			DELETE FROM workorder WHERE workorder_tag = _workorder_tag and workorder_scheduleDate > now() ;
-			call WORKORDER_create('create', _customerId, _userUUID, _workorderUUID, _workorder_locationUUID, _workorder_userUUID,
-								  _workorder_groupUUID, _workorder_assetUUID, _workorder_checklistUUID, _workorder_checklistHistoryUUID,
-								  _workorder_status, _workorder_type, _workorder_name, _workorder_number, _workorder_details, _workorder_actions,
-								  _workorder_priority, _workorder_dueDate, _workorder_completeDate, _workorder_scheduleDate, _workorder_rescheduleDate,
-								  _workorder_frequency, _workorder_frequencyScope, _wapj_asset_partUUID, _wapj_quantity, _daysToMaintain
-                                  );
+            IF (_workorder_tag IS NOT NULL) THEN
+                DELETE FROM workorder WHERE workorder_tag = _workorder_tag and workorder_scheduleDate > now() ;
+                call WORKORDER_create('create', _customerId, _userUUID, _workorderUUID, _workorder_locationUUID, _workorder_userUUID,
+                                    _workorder_groupUUID, _workorder_assetUUID, _workorder_checklistUUID, _workorder_checklistHistoryUUID,
+                                    _workorder_status, _workorder_type, _workorder_name, _workorder_number, _workorder_details, _workorder_actions,
+                                    _workorder_priority, _workorder_dueDate, _workorder_completeDate, _workorder_scheduleDate, _workorder_rescheduleDate,
+                                    _workorder_frequency, _workorder_frequencyScope, _wapj_asset_partUUID, _wapj_quantity, _daysToMaintain
+                                    );
 
-			LEAVE WORKORDER_workOrder;
+                LEAVE WORKORDER_workOrder;
+            END IF;
         END IF;
 
 		 set  @l_sql = CONCAT('update workorder set workorder_updatedTS=now(), workorder_updatedByUUID=', _userUUID);
@@ -1092,7 +1094,11 @@ ELSEIF(_action ='UPDATE') THEN
 			set @l_sql = CONCAT(@l_sql,',workorder_scheduleDate  = \'', _workorder_scheduledate ,'\'');
         END IF;
 
-		set @l_sql = CONCAT(@l_sql,' where workorderUUID = \'', _workorderUUID,'\';');
+        IF (_action = 'BATCH-UPDATE') THEN
+            set @l_sql = CONCAT(@l_sql,' where workorderUUID IN \('', _workorderUUID,')\';');
+        ELSE
+		    set @l_sql = CONCAT(@l_sql,' where workorderUUID = \'', _workorderUUID,'\';');
+        END IF;
 
         IF (_DEBUG=1) THEN select _action,@l_SQL; END IF;
